Moved client prometheus code to separate promclient package.
Nothing changed, except fixed IsWALDirAccessible (as WAL is actually a dir).
Otherwise just moved code.
