+++ cmd/thanos/sidecar.go
@@ -2,15 +2,10 @@ package main
 
 import (
 	"context"
-	"encoding/json"
-	"io/ioutil"
 	"math"
 	"net"
 	"net/http"
 	"net/url"
-	"os"
-	"path"
-	"path/filepath"
 	"sync"
 	"time"
 
@@ -19,6 +14,7 @@ import (
 	"github.com/improbable-eng/thanos/pkg/block/metadata"
 	"github.com/improbable-eng/thanos/pkg/cluster"
 	"github.com/improbable-eng/thanos/pkg/objstore/client"
+	"github.com/improbable-eng/thanos/pkg/promclient"
 	"github.com/improbable-eng/thanos/pkg/reloader"
 	"github.com/improbable-eng/thanos/pkg/runutil"
 	"github.com/improbable-eng/thanos/pkg/shipper"
@@ -31,7 +27,6 @@ import (
 	"github.com/prometheus/tsdb/labels"
 	"google.golang.org/grpc"
 	"gopkg.in/alecthomas/kingpin.v2"
-	"gopkg.in/yaml.v2"
 )
 
 func registerSidecar(m map[string]setupFunc, app *kingpin.Application, name string) {
@@ -253,7 +248,7 @@ func runSidecar(
 			}
 		}()
 
-		if err := isPrometheusDirAccesible(dataDir); err != nil {
+		if err := promclient.IsWALDirAccesible(dataDir); err != nil {
 			level.Error(logger).Log("err", err)
 		}
 
@@ -286,20 +281,6 @@ func runSidecar(
 	return nil
 }
 
-func isPrometheusDirAccesible(dir string) error {
-	const errMsg = "WAL directory is not accessible. Please ensure that the WAL / TSDB directory is accessible by the sidecar."
-	f, err := os.Stat(filepath.Join(dir, "wal"))
-	if err != nil {
-		return errors.Wrap(err, errMsg)
-	}
-
-	if !f.IsDir() {
-		return errors.New(errMsg)
-	}
-
-	return nil
-}
-
 type promMetadata struct {
 	promURL *url.URL
 
@@ -310,7 +291,7 @@ type promMetadata struct {
 }
 
 func (s *promMetadata) UpdateLabels(ctx context.Context, logger log.Logger) error {
-	elset, err := queryExternalLabels(ctx, logger, s.promURL)
+	elset, err := promclient.ExternalLabels(ctx, logger, s.promURL)
 	if err != nil {
 		return err
 	}
@@ -356,46 +337,4 @@ func (s *promMetadata) Timestamps() (mint int64, maxt int64) {
 	defer s.mtx.Unlock()
 
 	return s.mint, s.maxt
-}
-
-func queryExternalLabels(ctx context.Context, logger log.Logger, base *url.URL) (labels.Labels, error) {
-	u := *base
-	u.Path = path.Join(u.Path, "/api/v1/status/config")
-
-	req, err := http.NewRequest("GET", u.String(), nil)
-	if err != nil {
-		return nil, errors.Wrap(err, "create request")
-	}
-	resp, err := http.DefaultClient.Do(req.WithContext(ctx))
-	if err != nil {
-		return nil, errors.Wrapf(err, "request config against %s", u.String())
-	}
-	defer runutil.CloseWithLogOnErr(logger, resp.Body, "query body")
-
-	b, err := ioutil.ReadAll(resp.Body)
-	if err != nil {
-		return nil, errors.Errorf("failed to read body")
-	}
-
-	if resp.StatusCode != 200 {
-		return nil, errors.Errorf("got non-200 response code: %v, response: %v", resp.StatusCode, string(b))
-	}
-
-	var d struct {
-		Data struct {
-			YAML string `json:"yaml"`
-		} `json:"data"`
-	}
-	if err := json.Unmarshal(b, &d); err != nil {
-		return nil, errors.Wrapf(err, "unmarshal response: %v", string(b))
-	}
-	var cfg struct {
-		Global struct {
-			ExternalLabels map[string]string `yaml:"external_labels"`
-		} `yaml:"global"`
-	}
-	if err := yaml.Unmarshal([]byte(d.Data.YAML), &cfg); err != nil {
-		return nil, errors.Wrapf(err, "parse Prometheus config: %v", d.Data.YAML)
-	}
-	return labels.FromMap(cfg.Global.ExternalLabels), nil
 }
