+++ test/e2e/store_gateway_test.go
@@ -8,18 +8,17 @@ import (
 	"testing"
 	"time"
 
-	"github.com/improbable-eng/thanos/pkg/runutil"
-	"github.com/pkg/errors"
-	"github.com/prometheus/common/model"
-
 	"github.com/go-kit/kit/log"
 	"github.com/improbable-eng/thanos/pkg/objstore"
-	"github.com/prometheus/prometheus/pkg/timestamp"
-	"github.com/prometheus/tsdb/labels"
-
 	"github.com/improbable-eng/thanos/pkg/objstore/client"
 	"github.com/improbable-eng/thanos/pkg/objstore/s3"
+	"github.com/improbable-eng/thanos/pkg/promclient"
+	"github.com/improbable-eng/thanos/pkg/runutil"
 	"github.com/improbable-eng/thanos/pkg/testutil"
+	"github.com/pkg/errors"
+	"github.com/prometheus/common/model"
+	"github.com/prometheus/prometheus/pkg/timestamp"
+	"github.com/prometheus/tsdb/labels"
 	"gopkg.in/yaml.v2"
 )
 
@@ -95,7 +94,7 @@ func TestStoreGatewayQuery(t *testing.T) {
 		}
 
 		var err error
-		res, err = queryPrometheus(ctx, "http://"+queryHTTP(1), time.Now(), "{a=\"1\"}", false)
+		res, err = promclient.QueryInstant(ctx, nil, urlParse(t, "http://"+queryHTTP(1)), "{a=\"1\"}", time.Now(), false)
 		if err != nil {
 			return err
 		}
@@ -129,7 +128,7 @@ func TestStoreGatewayQuery(t *testing.T) {
 		}
 
 		var err error
-		res, err = queryPrometheus(ctx, "http://"+queryHTTP(1), time.Now(), "{a=\"1\"}", true)
+		res, err = promclient.QueryInstant(ctx, nil, urlParse(t, "http://"+queryHTTP(1)), "{a=\"1\"}", time.Now(), true)
 		if err != nil {
 			return err
 		}
